feat(gitlab): fix #7 added missing test

+++ issues-events.go
@@ -2,16 +2,16 @@ package main
 
 type IssuesEvents struct {
 	ObjectAttributes struct {
-		AssigneeID  int64       `json:"assignee_id"`
-		AuthorID    int64       `json:"author_id"`
+		AssigneeID  int         `json:"assignee_id"`
+		AuthorID    int         `json:"author_id"`
 		BranchName  interface{} `json:"branch_name"`
 		CreatedAt   string      `json:"created_at"`
 		Description string      `json:"description"`
-		ID          int64       `json:"id"`
-		Iid         int64       `json:"iid"`
-		MilestoneID int64       `json:"milestone_id"`
-		Position    int64       `json:"position"`
-		ProjectID   int64       `json:"project_id"`
+		ID          int         `json:"id"`
+		Iid         int         `json:"iid"`
+		MilestoneID int         `json:"milestone_id"`
+		Position    int         `json:"position"`
+		ProjectID   int         `json:"project_id"`
 		State       string      `json:"state"`
 		Title       string      `json:"title"`
 		UpdatedAt   string      `json:"updated_at"`
